sensor-simulator/requirements.txt
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
paho-mqtt==1.6.1
Pillow==9.1.1
prettytable==3.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.8.1
six==1.16.0
wcwidth==0.2.5
windows-curses==2.3.0
